Revert^2 "Add AddToDexClassloader JVMTI extension functions"
This reverts commit 799e536da9733ab638946f56e1ceb62d62cd3c81.

It seems that on some of our test devices the kernel does not have an
implementation for memfd_create. To work around this I added a basic
wrapper that will simulate memfd_create using temp files. This should
be sufficient for testing. All actual devices are expected to support
the memfd_create syscall natively.

Reason for revert: Implemented fallback for memfd_create
Bug: 132699522
Bug: 132914283
Test: ./test.py --host

Change-Id: I63b36464df24193fff27624c1e2350d65545ad1d
Signed-off-by: DennySPb <dennyspb@gmail.com>