Correctly delete all uses when building up the domination graph
We now consider the edge case where the dead instruction we wanted to
remove was used in a phi which is not directly following the block
we are processing right now. To fix this, we process all blocks
before trying to remove the instructions.

As a note, we remove catch phi uses right before removing the
instruction as they are related to the instruction itself and not
the block hierarchy.

Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b
Test: dex2oat compiling locally the apps mentioned in the bug
Bug: 239519319
Change-Id: I70d47891203ae118851a1f20a7cee21de305cd61
Signed-off-by: DennySPb <dennyspb@gmail.com>